simplify the base class for own TB6 integration tests
 - merge the 3 base classes into one: MultiBrowserTest + PrivateTB3Configuration + AbstractTB3Test -> AbstractTB6Test
 - remove unused methods
